Add patch for setting MTU on trunks.
With a large amount of traffic on IAX2 trunks, there is bad voice quality due to the fact that the
IAX2 trunking scheme depends on the Linux system to handle fragmentation of UDP packets. This is
not very efficient. This patch adds the ability to set an MTU size for *ALL* iAX2 trunks in
Asterisk 1.2. In the patch for svn trunk, we will add new features, like setting MTU per trunk.
In the case of Asterisk 1.2, we wanted a small clean patch to be able to get voice quality back
without changing a lot of source code.

The default behaviour is changed so that we send multiple UDP packets on IAX2 trunks where
the trunk packet size exceeds a standard Ethernet MTU minus overhead for PPTP and IPsec.
If you want to disable this behaviour, set trunkmtu=0 in the [general] section of iax.conf.

This patch is created by Voop AS in Norway.


git-svn-id: https://origsvn.digium.com/svn/asterisk/team/oej/iaxtrunkfix-1.2@44779 65c4cc65-6c06-0410-ace0-fbb531ad65f3