- Starting a cleanup proposal of AMI for 1.6 - Adding and changing manager events, all documented in CHANGES.moremanager - Based on 1.4 for testing, will be merged into trunk if accepted
This is not the BIG wonderful new manager, just a tiny cleanup so that a channel's name
is always reported as "Channel:" and nothing else.

Also, adding a few events to be able to follow channels better through the processing
and connecting a core call to a channel driver signalling channel (experimenting with
SIP of course).



git-svn-id: https://origsvn.digium.com/svn/asterisk/team/oej/moremanager@61757 65c4cc65-6c06-0410-ace0-fbb531ad65f3