Merged revisions 133941,133943,133945-133946 via svnmerge from file:///srv/subversion/repos/asterisk/trunk
........
r133941 | russell | 2008-07-26 09:46:13 -0500 (Sat, 26 Jul 2008) | 3 lines

Remove the code that decided when device state changes should be cached or not.
It is no longer needed.

........
r133943 | russell | 2008-07-26 09:57:50 -0500 (Sat, 26 Jul 2008) | 2 lines

Re-work comment about how device state changes are processed to be a bit more clear

........
r133945 | russell | 2008-07-26 10:15:14 -0500 (Sat, 26 Jul 2008) | 6 lines

ast_device_state() gets called in two different ways.  The first way is when
called from elsewhere in Asterisk to find the current state of a device.  In
that case, we want to use the cached value if it exists.  The other way is when
processing a device state change.  In that case, we do not want to check the
cache because returning the last known state is counter productive.

........
r133946 | russell | 2008-07-26 10:16:20 -0500 (Sat, 26 Jul 2008) | 1 line

actually use the cache_cache argument
........


git-svn-id: https://origsvn.digium.com/svn/asterisk/team/oej/astum@133964 65c4cc65-6c06-0410-ace0-fbb531ad65f3