Merged revisions 134649,134652 via svnmerge from file:///srv/subversion/repos/asterisk/branches/1.4
........
r134649 | tilghman | 2008-07-30 16:38:50 -0500 (Wed, 30 Jul 2008) | 4 lines

Qwell pointed out, via IRC, that the previous fix only worked when explicitly
set.  When nothing is set, and the option is implied, it breaks, because
configure sets the prefix to 'NONE'.  Fixing.

........
r134652 | murf | 2008-07-30 17:02:07 -0500 (Wed, 30 Jul 2008) | 19 lines

(closes issue #13197)
Reported by: pj

(closes issue #13051)
Reported by: pj

This patch substitutes commas in the expr 
supplied to the if () statement, as in
if ( expr ) ...

This solves both the bugs above, and makes
the source symmetric with switch statements,
which were earlier reported to need this sort
of treatment.

I tested this using the examples, both for
the compiler and at run time. Looks good.


........


git-svn-id: https://origsvn.digium.com/svn/asterisk/team/oej/sip-subscribe-ng@134687 65c4cc65-6c06-0410-ace0-fbb531ad65f3