Merged revisions 136168 via svnmerge from file:///srv/subversion/repos/asterisk/branches/1.4
........
r136168 | russell | 2008-08-06 13:58:37 -0500 (Wed, 06 Aug 2008) | 8 lines

Remove the use of --no-print-directory when compiling subdirectories.
This allows vim :make functionality to work properly when errors have
occurred in the build.  Without printing the directories, vim did not
know how to find the file that the error occurred in.

If the extra bit of build noise annoys anyone, just let me know, and
I'll make this optional.

........


git-svn-id: https://origsvn.digium.com/svn/asterisk/team/oej/sip-subscribe-ng@136182 65c4cc65-6c06-0410-ace0-fbb531ad65f3