Merged revisions 136544,136560 via svnmerge from file:///srv/subversion/repos/asterisk/branches/1.4
........
r136544 | sruffell | 2008-08-07 13:14:17 -0500 (Thu, 07 Aug 2008) | 11 lines

Updated codec_dahdi to use the new transcoder interface in the first DAHDI
release.  Codec dahdi no longer functions with the transcoder interface in
zaptel at this time (which the last zaptel release was 1.4.11).

NOTE: Still needs an update to the configure script to make sure that
codec_dahdi is only built if the new transcoder interface is present in the
drivers.

(Issue: DAHDI-42)


........
r136560 | kpfleming | 2008-08-07 13:25:31 -0500 (Thu, 07 Aug 2008) | 3 lines

change the required dependency for codec_dahdi to only be satisfied by DAHDI and not Zaptel, as the new transcoder interface is only in DAHDI


........


git-svn-id: https://origsvn.digium.com/svn/asterisk/team/oej/sipchanstats@136580 65c4cc65-6c06-0410-ace0-fbb531ad65f3