Merged revisions 137348 via svnmerge from file:///srv/subversion/repos/asterisk/branches/1.4
........
r137348 | seanbright | 2008-08-12 19:10:35 -0500 (Tue, 12 Aug 2008) | 8 lines

Bring cdr_tds in line with the other CDR backends and have it try to store
CDR(userfield) if it is set.  The new behavior is to check for the userfield
column on module load, and if it exists, we will store CDR(userfield) when
CDRs are written.  A similar patch already went into trunk and 1.6.0.

(closes issue #13290)
Reported by: falves11

........


git-svn-id: https://origsvn.digium.com/svn/asterisk/team/oej/cancel_elsewhere_1.4@137378 65c4cc65-6c06-0410-ace0-fbb531ad65f3