Merged revisions 138119 via svnmerge from file:///srv/subversion/repos/asterisk/branches/1.4
........
r138119 | jpeeler | 2008-08-15 14:21:51 -0500 (Fri, 15 Aug 2008) | 4 lines

Fixes the dahdi restart functionality. Dahdi restart allows one to restart all DAHDI channels, even if they are currently in use. This is different from unloading and then loading the module since unloading requires the use count to be zero. Reloading the module is different in that the signalling is not changed from what it was originally configured. Also, this fixes not closing all the file descriptors for D-channels upon module unload (which would prevent loading the module afterwards).

(closes issue #11017)

........


git-svn-id: https://origsvn.digium.com/svn/asterisk/team/oej/cancel_elsewhere_1.4@138125 65c4cc65-6c06-0410-ace0-fbb531ad65f3