Merged revisions 141217 via svnmerge from file:///srv/subversion/repos/asterisk/branches/1.4
........
r141217 | mmichelson | 2008-09-05 11:00:24 -0500 (Fri, 05 Sep 2008) | 14 lines

Commit 140417 had a logic flaw in it which
caused port 5060 to always be used when dialing
a peer if no explicit port was specified. This
broke the behavior of implicitly using the port
from which the peer registered if no port is
specified. This commit fixes the logic flaw.

(closes issue #13424)
Reported by: mdu113
Patches:
      13424.patch uploaded by putnopvut (license 60)
Tested by: mdu113


........


git-svn-id: https://origsvn.digium.com/svn/asterisk/team/oej/sip-subscribe-ng@141255 65c4cc65-6c06-0410-ace0-fbb531ad65f3