Merged revisions 142354,142358 via svnmerge from file:///srv/subversion/repos/asterisk/branches/1.4
........
r142354 | russell | 2008-09-10 11:39:53 -0500 (Wed, 10 Sep 2008) | 7 lines

It is a normal situation that a task gets put in the scheduler that should run
as soon as possible.  Accept "0" as an acceptable time to run, and also treat
negative as "run now", and don't print a debug message about it.

(inspired by a message asking about the "request to schedule in the past"
 debug message on the -dev list)

........
r142358 | tilghman | 2008-09-10 11:54:29 -0500 (Wed, 10 Sep 2008) | 2 lines

Publish new extra sounds version.

........


git-svn-id: https://origsvn.digium.com/svn/asterisk/team/oej/minivoicemail-1.4@142393 65c4cc65-6c06-0410-ace0-fbb531ad65f3