Merged revisions 143674 via svnmerge from file:///srv/subversion/repos/asterisk/branches/1.4
........
r143674 | murf | 2008-09-19 16:07:05 -0500 (Fri, 19 Sep 2008) | 1 line

This fix comes from a debugging session on a test box that has been getting hung channels when the mssqlserver bounces. All the connections then become invalid, and must be reconnected. The cdr_odbc backend had code to do it, but depended on re-establishing the connection, but re-using the STMT that had been built. By trial and error, we determined that the STMT could not be re-used after the connection was re-established. and must be rebuilt. These changes accomplish this.
........


git-svn-id: https://origsvn.digium.com/svn/asterisk/team/oej/minivoicemail-1.4@143687 65c4cc65-6c06-0410-ace0-fbb531ad65f3