Merged revisions 143964 via svnmerge from file:///srv/subversion/repos/asterisk/branches/1.4
........
r143964 | murf | 2008-09-23 09:22:10 -0500 (Tue, 23 Sep 2008) | 21 lines

In at least one machine, we noted that the timestr
was not getting set in the STMT; it was coming out,
usually, as binary garbage to an mssql server.
These changes fixed the problem. The only thing
I can venture forth as a guess, is that the pointer
is being stored in the interface, not a copy of the
string. Because we ripped the build process into a 
subroutine, the timestr became a temp. stack variable,
and between the time the STMT got built and the
time it was executed on the server, the string being
pointed to was damaged. At any rate, even if this
theory is false, and some mechanism was at fault,
this fix worked reliably where it didn't before.

Why this bug didn't bite last week, I have no idea.
This change basically defines the timestr buffer
in the calling function, extending the life of the
buffer to cover both the STMT's building and
processing to the server.


........


git-svn-id: https://origsvn.digium.com/svn/asterisk/team/oej/cancel_elsewhere_1.4@143973 65c4cc65-6c06-0410-ace0-fbb531ad65f3