Merged revisions 144758 via svnmerge from file:///srv/subversion/repos/asterisk/branches/1.4
........
r144758 | mmichelson | 2008-09-26 17:14:59 -0500 (Fri, 26 Sep 2008) | 8 lines

This patch was applied to 1.4 but it completely
does not apply since the "found" pointer is not
passed in to this function. If this is going to
be backported, it needs to be done differently or
a deeper backport needs to be done.

Edit: This commit reverts commit number 144677.

........


git-svn-id: https://origsvn.digium.com/svn/asterisk/team/oej/sip-subscribe-ng@144799 65c4cc65-6c06-0410-ace0-fbb531ad65f3