Merged revisions 158600,158603,158629 via svnmerge from file:///srv/subversion/repos/asterisk/branches/1.4
........
r158600 | tilghman | 2008-11-21 17:07:46 -0600 (Fri, 21 Nov 2008) | 5 lines

The passed extension may not be the same in the list as the current entry,
because we strip spaces when copying the extension into the structure.
Therefore, use the copied item to place the item into the list.
(found by lmadsen on -dev, fixed by me)

........
r158603 | murf | 2008-11-21 17:14:50 -0600 (Fri, 21 Nov 2008) | 11 lines

In reference to the fix made for 13871, I was
merging the fix into 1.6.0 and realized I missed
the code in the h-exten block, and didn't catch it
because my test case had the h-exten commented out.

So, this corrects the code I missed, as a 
preventative against another crash report.
Tested with the h-exten defined, all is well.



........
r158629 | jpeeler | 2008-11-21 18:04:36 -0600 (Fri, 21 Nov 2008) | 5 lines

(closes issue #13786)
Reported by: tzafrir

When compiling against Zaptel dahdi_compat will now only define all the DAHDI defines if the Zaptel define is present. Also, there is no such thing as DAHDI_PRI.

........


git-svn-id: https://origsvn.digium.com/svn/asterisk/team/oej/fixtoheader-1.4@158637 65c4cc65-6c06-0410-ace0-fbb531ad65f3