Merged revisions 160480 via svnmerge from file:///srv/subversion/repos/asterisk/branches/1.4
........
r160480 | tilghman | 2008-12-03 08:09:35 -0600 (Wed, 03 Dec 2008) | 7 lines

Jon Bonilla (Manwe) pointed out on the -dev list:
"I guess that having only ip-phones in mind is not a good approach. Since it is
possible to have a sip proxy connected to asterisk we could receive a 407
(unauthorized) or 483 (too many hops) as response and dialog ending would not be
a good behavior."
So modified.

........


git-svn-id: https://origsvn.digium.com/svn/asterisk/team/oej/fixtoheader-1.4@160517 65c4cc65-6c06-0410-ace0-fbb531ad65f3