Merged revisions 160943 via svnmerge from file:///srv/subversion/repos/asterisk/branches/1.4
........
r160943 | mmichelson | 2008-12-04 10:44:18 -0600 (Thu, 04 Dec 2008) | 15 lines

Fix a callerid parsing issue. If someone formatted callerid like the
following: "name <number>" (including the quotation marks), then the parts
would be parsed as 

name: "name
number: number

This is because the closing quotation mark was not discovered since the number
and everything after was parsed out of the string earlier. Now, there is a check
to see if the closing quote occurs after the number, so that we can know if we
should strip off the opening quote on the name.

Closes AST-158


........


git-svn-id: https://origsvn.digium.com/svn/asterisk/team/oej/sipchanstats@160983 65c4cc65-6c06-0410-ace0-fbb531ad65f3