Merged revisions 161948 via svnmerge from file:///srv/subversion/repos/asterisk/branches/1.4
........
r161948 | russell | 2008-12-09 08:52:25 -0600 (Tue, 09 Dec 2008) | 15 lines

Fix a problem with GROUP() settings on a masquerade.

The previous code carried over group settings from the old channel to the new
one.  However, it did nothing with the group settings that were already on the
new channel.  This patch removes all group settings that already existed on the
new channel.

I have a more complicated version of this patch which addresses only the most
blatant problem with this, which is that a channel can end up with multiple
group settings in the same category.  However, I could not think of a use case
for keeping any of the group settings from the old channel, so I went this route
for now.

(closes AST-152)

........


git-svn-id: https://origsvn.digium.com/svn/asterisk/team/oej/iaxtrunkfix-1.4@161986 65c4cc65-6c06-0410-ace0-fbb531ad65f3