Merged revisions 164672 via svnmerge from file:///srv/subversion/repos/asterisk/branches/1.4
........
r164672 | russell | 2008-12-16 09:56:37 -0600 (Tue, 16 Dec 2008) | 11 lines

Fix a memory leak related to the use of the "setvar" configuration option.

The problem was that these variables were being appended to the list of vars
on the sip_pvt every time a re-registration or re-subscription came in.
Since it's just a waste of memory to put them there unless the request was an
INVITE, then the fix is to check the request type before copying the vars.

(closes issue #14037)
Reported by: marvinek
Tested by: russell

........


git-svn-id: https://origsvn.digium.com/svn/asterisk/team/oej/iaxtrunkfix-1.4@164708 65c4cc65-6c06-0410-ace0-fbb531ad65f3