Appleraisin for trunk
Experimenting with porting astdb to realtime, so you can store all
astdb operations in any storage engine you want. This helps with
old or stubborn applications that still wants to use a lot of
astdb operations from the AMI, cli and/or dialplan.

Seems like the trunk realtime has the needed operations for doing
this, something that does not exist in the realtime for 1.4.


git-svn-id: https://origsvn.digium.com/svn/asterisk/team/oej/codename-appleraisin-trunk@176553 65c4cc65-6c06-0410-ace0-fbb531ad65f3