Test branch for patches hidden in several branches - all based on Asterisk 1.4
- RTCP improvements from pinefrog-1.4
- "Sip show chanstats" 
- pinequality-* giving you the manager "sipchannel" event

This branch is now open for testing and I need feedback. Among the improvements

- Manager QoS events during a call and after a call
- Improved RTCP - rtcp now works for p2p bridge in RTP, which means that we will get 
  RTCP for many, many more sip calls
- RTCP over NAT - if Asterisk is behind NAT, we will now kick-start RTCP from the remote
  end by sending a first "emtpy" RTCP packet to open a NAT port. 
- QoS reports to realtime storage after each call - one report per call leg
  (The amount of data and the names will change)

The reason I store data in realtime, is that the CDR is usually gone or frozen at the time 
that we freeze the RTP channels and get the last QoS data. The QoS reports can't thus 
be included in CDR, you have to merge it in automatically later in your database.

There's still a lot to do, but please test it so I get some sort of feedback.

For testing, don't forget to run the "rtcp debug" cli command so you can see what's
going on in the RTCP channel.

FAQ
---
Yes, this work will be ported to trunk and hopefully merged soon. 
No, I have no reason or funding to adapt it to 1.6.x at this point.
No, the RTPAUDIOQOS is not changed at all. YOu might get more data now though.

-------
This work is funded to 20% by companies in the community. If you want to cover the
80% that's still not funded, please contact me off-list (oej@edvina.net).




git-svn-id: https://origsvn.digium.com/svn/asterisk/team/oej/pinefrog-deluxe-rtcp-test@241850 65c4cc65-6c06-0410-ace0-fbb531ad65f3