This is the first round of fixes for one of the problems mentioned in bug 17055, that wasn't split off into another bug report, namely, that "also while the announce is being played the call can not be answered by an agent"
What is happening is that each sound file is played in the current thread, and a XXX_wait()
call (usually ast_waitstream()) is made, to process the sound into the channel. The result is that app_queue is 
frozen until the sound is finished playing. This can get cumbersome!

I changed it so no wait() calls are made. I use a channel datastore to sort of 
mark the channel that it's in this special mode. THe datastore has a struct that
contains all the data we could desire, including the pointer to an eof-handler on the
filestream. That's the key, enabling a handler to be run when a sound file is finished.

The tricky part about the eof-handler for filestreams, is where to call it! I found about
the only practical spot, in main/file.c, in the func read_frame. Luckily, it's up one
level from all the format read funcs, and gets a null from those funcs when they hit eof.

So, we do the ast_streamfile() call, and fire off ast_autoservice_start() to push the audio
out in a different thread. When the file finishes, the handler is fired up to shut off 
autoservice, start music on hold, or whatever else would need to be done sound-wise at that point.

So, basically, if MOH is turned on, app_queue is fundamentally an MOH player, with interruptions
to play sound files. 

To really complicate the affair, it plays sequences of files including doing ast_say_number() calls.
I queue the other files up in the datastore in a list, and use the add_playfile func in app_queue to
do this.

But for the say_number stuff, I needed to insinuate this capability into the wait_file func, and to do that,
I had to remove all the waitstream calls, and make the whole say.c use the same basic func to play (and maybe wait for) all files.
Really, this should have been done a while ago. It was the same two or three sets of code repeated over a hundred
times, with several bugs introduced. I'd like to think the new wait_file func (and wait_file_full) reduce the size
of the file and make it more stable. I've marked a set of waitstream calls that are called between say_number calls,
because really, they should not be necessary. And, here and there, folks forgot to call waitstrem after ast_streamfile, which
should have ended up skipping the playing of that file. This kind of thing is fixed. say.c should no longer use
ast_waitstream, ast_waitstream_full, ast_streafile, or ast_play_and_wait, but rather, just wait_file or wait_file_full,
and those guys get to use the lower level calls.

In app_queue, the play_file func will interrupt any other playing files, including moh. The handler should get MOH going
again after the file finishes playing.

the add_playfile func will add the file to the list in the datastore. The eof-handler will begin the streaming of the 
first file in the list (if there), and remove it from the list. All the say_number calls will generally just end up
doing add_playfile calls for reading off the number.

I can't guarantee that this stuff compiles, but even if it does, I *can* guarantee that it won't work. For now.
I still have to refactor the sound related code that is done after the files are played, and make sure those
operations will be done in the eof-handler.  All and all, it's almost done.

  


git-svn-id: https://origsvn.digium.com/svn/asterisk/team/murf/pinequeue@279343 65c4cc65-6c06-0410-ace0-fbb531ad65f3