Made several fixes and rearrangements;
first of all, coalesced the play_file and add_playfile into play_file. All
we need is one func, because, with all its previous waiting,
the behavior of app_queue is to queue the playing of files in time. So,
no pre-emptive behavior in this code either. So, one func
will work fine. This required some changes in say.c also.

Next, I got rid of code no longer necc. to restart MOH here and there.

I clean up the channel and get rid of my stuff on exiting app_queue.

The only big issue left is what to do with the autoservice behavior on 
announcing to the peer... more to come soon.




git-svn-id: https://origsvn.digium.com/svn/asterisk/team/murf/pinequeue@283769 65c4cc65-6c06-0410-ace0-fbb531ad65f3