Wonderful Oppurtunity - never heard of before! ==============================================
Review this branch - and you'll get the chance to win a 
FREE license of Asterisk - the leading Open Source PBX! 
If you do it before the end of the month, I'll throw in 
a FREE license of Kamailio -the super duper SIP server - 
as well!!!

https://reviewboard.asterisk.org/r/923/

Thank you for your attention. This offer is only valid within
Sweden until Oct 1 2010 for people who like Abba and are not
employees of Edvina AB or it's partners. You are only allowed
to do one review per person and will only get one ticket to
win the free license per individual. Reviewbots will not
be allowed to participate in this competition. The Swedish
requirement does not apply to people working in Huntsville, AL, USA
and Vence, France, EU. You may review in Swedish, English or
any dialect understood in the northern forestes of Orsa.


git-svn-id: https://origsvn.digium.com/svn/asterisk/team/group/pine-cdr-pgsql-status-trunk@288486 65c4cc65-6c06-0410-ace0-fbb531ad65f3