Adding a new structure for sip2cause.
By adding a dynamic structure instead of hardcoding the conversion, we open up
for a new configuration that can change this.

Trying to make up my mind if I just want to add a new dynamic table that is
checked first, or allow editing of this one. 

But first, try to figure out if this works.


git-svn-id: https://origsvn.digium.com/svn/asterisk/team/oej/earl-grey-sip2cause-configurable-1.8@376755 65c4cc65-6c06-0410-ace0-fbb531ad65f3