Fooling around with packet loss concealment
In the RTP channel we actually notice that there are packets missing in the incoming 
RTP stream. What would happen if we insert a copy of the previous packet into the stream
at that point. Well, we will introduce jitter as we're sending two frames at the same time, 
but hopefully the endpoints jitter buffer will sort that out.

Other functions, like recordings, will actually get 50 frames with 20 ms audio (or the needed amount)
per second and we will have more complete streams.

For short calls, this really doesn't make much of a difference, but we have situations with
three or four hour calls and the ackumulated packet loss seems to cause bad issues with recordings
and maybe other stuff as well. We're also loosing synch between inbound and outbound streams.


git-svn-id: https://origsvn.digium.com/svn/asterisk/team/group/pinefool-poor-mans-plc-1.4@383403 65c4cc65-6c06-0410-ace0-fbb531ad65f3