Add +SIP.INSTANCE support to outbound registrations and subscriptions in SIP
This is a small identifier of a device or a server that is unique. Adding it to inbound registrations,
we can decide to remove registrations that have the same instance ID as the incoming new Contact.
We should also be able to implement that we deny registrations from another device as long as we
have a valid registration from one instance ID (one specific device).

This identifier is used in many new SIP standards, like SIP outbound, GRUU and GIN.


git-svn-id: https://origsvn.digium.com/svn/asterisk/team/oej/pine-instance-uuid-1.8@386614 65c4cc65-6c06-0410-ace0-fbb531ad65f3