Fix memory leak while loading priority modules
When we load a module with the LOAD_PRIORITY flag, we remove its entry from
the load order list. Unfortunately, we don't free the memory associated with
entry in the list. This patch corrects that and properly frees the memory for
the module in the list.
........

Merged revisions 391489 from file:///srv/subversion/repos/asterisk/branches/1.8


git-svn-id: https://origsvn.digium.com/svn/asterisk/team/oej/assam-early-media-focus-1.8@391497 65c4cc65-6c06-0410-ace0-fbb531ad65f3