test_voicemail_api: fix warning found by gcc-4.8
The voicemail_api test had code like strncmp(a, b, sizeof(a)), but a was a
char pointer, instead of a literal or char array. This meant that sizeof was
the size of the pointer, not the length of the string.

Since the string is in a stringfield and should be null terminated, I just
changed it to a plain strcmp.
........

Merged revisions 394050 from file:///srv/subversion/repos/asterisk/trunk
........

Merged revisions 394056 from file:///srv/subversion/repos/asterisk/team/rmudgett/bridge_phase


git-svn-id: https://origsvn.digium.com/svn/asterisk/team/rmudgett/bridge_tasks@394057 65c4cc65-6c06-0410-ace0-fbb531ad65f3