Fix invalid access to disposed memory in main/data unit test
It is not safe to iterate over a macro'd list of ao2 objects, deref them such
that the item's destructor is called, and leave them in the list. The list
macro to iterate over items requires the item to be a valid allocated object
in order to proceed to the next item; with MALLOC_DEBUG on the corruption of
the linked list is caught in the crash.

This patch fixes the invalid access to free'd memory by removing the ao2 item
from the list before de-refing it.
........

Merged revisions 396915 from file:///srv/subversion/repos/asterisk/trunk
........

Merged revisions 396918 from file:///srv/subversion/repos/asterisk/team/rmudgett/bridge_phase


git-svn-id: https://origsvn.digium.com/svn/asterisk/team/rmudgett/bridge_tasks@396919 65c4cc65-6c06-0410-ace0-fbb531ad65f3