Add a res_bucket_sounds module which implements a 'sounds' URI scheme.
This scheme is a virtual view of the sounds available on the system. Buckets
are created which contain the languages as well as sub-directories. Files are
created which allow getting a sound at the format level and non-format level.
Metadata is placed on files which contain useful information (such as languages,
formats, and description). There is also a top level root bucket for getting at
everything.

Reloads are done in an atomic fashion and will fail if any critical error occurs.


git-svn-id: https://origsvn.digium.com/svn/asterisk/team/file/res_bucket_sounds@398495 65c4cc65-6c06-0410-ace0-fbb531ad65f3