Update CDR Unit tests to reflect container changes in r398579
When a channel joins a multi-party bridge, the ordering of the CDRs that is
created is determined by the ordering of the channels who happen to be in that
bridge. When r398579 changed the number of buckets in the container to
something sensible, it changed the ordering that the CDRs was created in,
causing one of the multiparty tests to fail. This fixes the test with the
now expected ordering.
........

Merged revisions 398628 from http://svn.asterisk.org/svn/asterisk/branches/12
........

Merged revisions 398629 from file:///srv/subversion/repos/asterisk/trunk


git-svn-id: https://origsvn.digium.com/svn/asterisk/team/rmudgett/ao2_enhancements@398632 65c4cc65-6c06-0410-ace0-fbb531ad65f3