Introducing working shadow peers.
For each entry in the SRV list, we create a shadow peer for matching of IP/port on incoming calls. When we match, we simply
redirect to the master peer to get call counters and other things working correctly.

When reloading, we remove all the shadow peers and create a new list. Will try to make a timer for TTL expiry so that we
renew automatically when TTL for SRV expire too.


git-svn-id: https://origsvn.digium.com/svn/asterisk/team/oej/pgtips-srv-and-outbound-stuff-1.8@403235 65c4cc65-6c06-0410-ace0-fbb531ad65f3