After a discussion with Liman, add a todo to make shadow peers configurable.
Need a meeting on whether we should have bullet #13 or just skip from #12 to #14. A fika and a 
typical Swedish meeting is what's needed to make a decision here. Who can join, where do we discuss
date and time? Do we need to stream it over the Internet? I guess we have to do this properly.

I guess we will need a pre-meeting to discuss this meeting. Of course the pre-meeting for planning
the real meeting will need fika too.

Fika is very important. http://en.wikipedia.org/wiki/Fika_(coffee_break)


git-svn-id: https://origsvn.digium.com/svn/asterisk/team/oej/pgtips-srv-and-outbound-stuff-1.8@403237 65c4cc65-6c06-0410-ace0-fbb531ad65f3