res_pjsip_pubsub: Ensure dialog manipulation happens on proper thread.
When destroying a subscription we remove the serializer from its dialog
and decrease its reference count. Depending on which thread dropped the
subscription reference count to 0 it was possible for this to occur in
a thread where it is not possible.

(closes issue ASTERISK-22952)
Reported by: Matt Jordan
........

Merged revisions 404553 from file:///srv/subversion/repos/asterisk/branches/12


git-svn-id: https://origsvn.digium.com/svn/asterisk/team/mjordan/12-memorypool@404556 65c4cc65-6c06-0410-ace0-fbb531ad65f3