cdrs: Check for applications to lock onto during dial begin handling
This patch brings CDR processing further in line with r407085. During some dial
operations, the application would not be locked to the Dial application and
would instead continue to show the previously known application. In particular,
this would occur when a Parked call would time out. This was due to a previous
snapshot already locking the application to Park - processing this in a Dial
Begin allows the Dial application to reassert its rightful place.

(CDRs. Ugh.)

But hooray for the Parked Call tests for catching this in the Asterisk Test
Suite.
........

Merged revisions 407166 from http://svn.asterisk.org/svn/asterisk/branches/12
........

Merged revisions 407169 from file:///srv/subversion/repos/asterisk/trunk


git-svn-id: https://origsvn.digium.com/svn/asterisk/team/rmudgett/bridge_tasks@407176 65c4cc65-6c06-0410-ace0-fbb531ad65f3