config: Fix inverted test
The test of the result of the stat() call was inverted such that its
output was only used if the call failed. This inverts the test so that
the output of stat() is used correctly. This was causing full reloads
on unchanged files.

(closes issue ASTERISK-23383)
Reported by: David Woolley
........

Merged revisions 409916 from file:///srv/subversion/repos/asterisk/branches/1.8


git-svn-id: https://origsvn.digium.com/svn/asterisk/team/oej/oolong-path-support-1.8@409947 65c4cc65-6c06-0410-ace0-fbb531ad65f3