Make sure we use the proper address
Now, this is a situation where the IP address in the registry object does not
match the one actually used for the REGISTER transaction. We propably should
update the registry object as well. I am surprised that we look up the DNS
address so many times for a single REGISTER. With round robin DNS, the address
keeps changing, which cause bad things to happen.


git-svn-id: https://origsvn.digium.com/svn/asterisk/team/oej/register-call-dependency-1.8@422235 65c4cc65-6c06-0410-ace0-fbb531ad65f3