Blocked revisions 428104
........
tests/test_cel: Fix CEL unit tests

This is a backport of the test_cel portion of r427870, which was not applied to
the 12 branch. This fixes the compilation issues with the CEL unit tests
introduced by the API changes needed to fix publication of blind transfer
messages.
........

Merged revisions 428105 from file:///srv/subversion/repos/asterisk/branches/13


git-svn-id: https://origsvn.digium.com/svn/asterisk/team/mjordan/13-swagger@428110 65c4cc65-6c06-0410-ace0-fbb531ad65f3