Allow for transferer to retry when dialing an invalid extension.
This allows for a configurable number of attempts for a transferer
to dial an extension to transfer the call to. For Asterisk 13, the
default values are such that upgrading between versions will not
cause a behaivour change. For trunk, though, the defaults will be
changed to be more user-friendly.

Review: https://reviewboard.asterisk.org/r/4167
........

Merged revisions 428145 from file:///srv/subversion/repos/asterisk/branches/13


git-svn-id: https://origsvn.digium.com/svn/asterisk/team/mjordan/13-swagger@428157 65c4cc65-6c06-0410-ace0-fbb531ad65f3