apps/app_mixmonitor: Move Test Event for MIXMONITOR_END to after it finishes
The Test Event for MIXMONITOR_END - which signals that a MixMonitor has
completed - technically fired before the filestream was closed. If a test
used this to trigger a condition to verify that the file was written, it
could result in a race condition where the file size would not be what the
test expected.

Luckily, no tests were using this (although they should have been). Since the
test event needed to be moved after the point where the MixMonitor autochan has
been destroyed, the test event no longer emits the channel name. Luckily,
nothing needs it.
........

Merged revisions 431788 from file:///srv/subversion/repos/asterisk/branches/11


git-svn-id: https://origsvn.digium.com/svn/asterisk/team/oej/chocolate-video-congestion-11@431804 65c4cc65-6c06-0410-ace0-fbb531ad65f3