res_ari_channels: Return a 404 response when a requested channel variable does not exist.
This change makes it so that if a channel variable is requested and it does not exist
a 404 response will be returned instead of an allocation failed response. This makes
it easier to debug and figure out what is going on for a user.

ASTERISK-24677 #close
Reported by: Joshua Colp
........

Merged revisions 432154 from http://svn.asterisk.org/svn/asterisk/branches/13
........

Merged revisions 432155 from file:///srv/subversion/repos/asterisk/trunk


git-svn-id: https://origsvn.digium.com/svn/asterisk/team/group/generic-tts@432158 65c4cc65-6c06-0410-ace0-fbb531ad65f3