alemebic scripts: endpoint identifier order option
The script was added in 13, but when committed to trunk it caused a branch to
occur due to some trunk only alemebic changes. This fixes it so that the new
'add_pjsip_endpoint_identifier_order script points to the correct down revision.
........

Merged revisions 433152 from file:///srv/subversion/repos/asterisk/trunk


git-svn-id: https://origsvn.digium.com/svn/asterisk/team/group/generic-tts@433156 65c4cc65-6c06-0410-ace0-fbb531ad65f3