Add basic DNS NAPTR record handling.
In order to be able to allocate NAPTR records correctly, I had
to make an additional change to the ast_dns_result structure to
include the size of the DNS answer.

There is an incomplete test in res_resolver_unbound.c that was
being used simply to ensure that NAPTR records were being returned
and parsed as expected. The test will be expanded to actually
check values in the near future.

Also, it's probably best for your sanity if you don't look too
deeply at the NAPTR record allocation function.



git-svn-id: https://origsvn.digium.com/svn/asterisk/team/group/dns_naptr@433196 65c4cc65-6c06-0410-ace0-fbb531ad65f3