Add sorting to DNS NAPTR records.
I modeled this after the method being done in the SRV branch. That is,
I have an internal function that is defined in dns_naptr.c to sort the
records. I opted to sort records after they are collected in order to
make the code more uniform: SRV and NAPTR work the same way.

In order to test this, I added some more NAPTR records to the nominal
NAPTR resolution test and gave it a shot. This pointed to an error where
I needed to cast to unsigned char when getting the order and preference
from the NAPTR record.



git-svn-id: https://origsvn.digium.com/svn/asterisk/team/group/dns_naptr@433314 65c4cc65-6c06-0410-ace0-fbb531ad65f3