Backport revision 429223 from Asterisk 13
The bug fixed by that patch exists in Asterisk 11 as
well, so the fix should be applied there.

When connecting to a remote Asterisk console, the buffer
used to read the initial hostname/pid/version from the
main Asterisk process did not ensure the input was
NULL-terminated, and the buffer was small for certain
use cases. This patch expands the buffer to be 256
bytes and ensures the input it reads is NULL-terminated.

ASTERISK-21854 #close
Reported by klaus3000
........

Merged revisions 433919 from file:///srv/subversion/repos/asterisk/branches/11


git-svn-id: https://origsvn.digium.com/svn/asterisk/team/oej/chocolate-video-congestion-11@433920 65c4cc65-6c06-0410-ace0-fbb531ad65f3